+++ R/10_obs_hysplit.R
@@ -30,7 +30,13 @@
 #' @param top_model_domain altitude above ground level (m).
 #' @param met meteorological models to be used.
 #' @param nmet Number of days for the meteorological files. Default is
-#' number of days in duration plus two days
+#' number of days in duration plus two days.
+#' nmet is the number of simultaneous input meteorological files. For instance,
+#' 11 means that for each meteorological grid, 11 files are expected. Usually,
+#' the files are daily. Note that the same number of files are required for
+#' each grid in this approach. Hysplit expects something like 2 11,
+#' which means 2 meteorological grids with 11 files each. The number
+#' 2 comes from the length of met files.
 #' @param metpath paths for each meteorological model output.
 #' @param ngases Default 1.
 #' @param gas default "Foot".
@@ -83,11 +89,11 @@ obs_hysplit_control <- function(df,
                                 duration = -240,
                                 vertical_motion = 5,
                                 top_model_domain = 20000,
-                                met = c("nams", "gfs0p25", "era5"),
+                                met = c("hrrr", "nams", "gfs0p25"),
                                 nmet =  abs(duration/24) + 1, # 10 days plus 1, defaultdays
-                                metpath = c("/work/noaa/lpdm/metfiles/nams/",
-                                            "/work/noaa/lpdm/metfiles/gfs0p25/",
-                                            "/work/noaa/lpdm/metfiles/era5/arl/"),
+                                metpath = c("/work/noaa/lpdm/metfiles/hrrr/",
+                                            "/work/noaa/lpdm/metfiles/nams/",
+                                            "/work/noaa/lpdm/metfiles/gfs0p25/"),
                                 ngases = 1,
                                 gas = "Foot",
                                 emissions_rate = 0,
@@ -178,7 +184,7 @@ obs_hysplit_control <- function(df,
 
   sink(control)
 
-  cat(substr(yr, 3, 4)) #I need to confirm
+  cat(substr(yr, 3, 4))
   cat(" ")
 
   cat(sprintf(mo, fmt = '%02d'))
@@ -263,7 +269,7 @@ obs_hysplit_control <- function(df,
       }
     }
 
-    if(metx == "era5") {
+    if(metx == "hrrr") {
       for(i in  1:nmet){
         hyd <- as.Date(hydate - i)
 
@@ -276,6 +282,25 @@ obs_hysplit_control <- function(df,
 
         cat(
           paste0(strftime(hyd, format = "%Y%m%d"),
+                 "_hrrr"))
+        cat("\n")
+
+      }
+    }
+
+    if(metx == "era5") {
+      for(i in  1:nmet){
+        hyd <- as.Date(hydate - i)
+
+        cat(
+          paste0(metpath[j],
+                 data.table::year(hyd),
+                 "/"))
+        cat("\n")
+
+
+        cat(
+          paste0(strftime(hyd, format = "ERA5_%Y%m%d"),
                  ".ARL"))
         cat("\n")